+++ docs/src/examples/QasDialog/ExWithActions.vue
@@ -17,10 +17,7 @@ export default {
   computed: {
     dialogProps () {
       return {
-        card: {
-          description: 'Cerca elétrica perto do corrimão risco iminente de choque. Código da ordem de serviço: 159488 CRM ID: 1179512'
-        },
-
+        description: 'Cerca elétrica perto do corrimão risco iminente de choque. Código da ordem de serviço: 159488 CRM ID: 1179512',
         ok: { onClick: () => alert('clicado') }
       }
     }